refactor(kernel-utils): inline driveLift into sheafify and remove drive.ts
